installed typescript moved app.js into src, moved folders in src into db, renamed db to database, and fixed filepath issues arising from the rearrangement changed file extensions to .ts for all src files created tsconfig.json file and added node type to it so that it would work with require statements added and changed scripts to package.json so that the program could be cleaned, built, and ran with just one command added rimraf package to package.json so that the build folder's contents could be deleted before being replaced by a tsc command edited the README.md file: removed outdated information and added instructions for setting up a development environment and running the project
