+++ src/app.ts
@@ -1,8 +1,9 @@
+(() => {
 const { ApolloServer, gql } = require('apollo-server');
 const { connect, model, Schema } = require('mongoose');
-const url = require('./src/db/config');
-const typeDefs = require('./src/typeDefs');
-const resolvers = require('./src/resolvers');
+const url = require('./database/config');
+const typeDefs = require('./typeDefs');
+const resolvers = require('./resolvers');
 
 const port = process.env.PORT || 3000;
 
@@ -19,3 +20,4 @@ connect(url)
     .then((result) => {
         console.log("Server running at " + result.url);
     })
+})();
