+++ src/database/config.ts
@@ -1,5 +1,7 @@
+(() => {
 require('dotenv').config();
 
 const url = process.env.DATABASE_URL;
 
 module.exports = url;
+})();
