+++ src/database/users/users-model.ts
@@ -1,3 +1,4 @@
+(() => {
 const { model, Schema } = require('mongoose');
 
 const userSchema = new Schema({
@@ -9,3 +10,4 @@ const userSchema = new Schema({
 const User = model('User', userSchema);
 
 module.exports = User;
+})();
