+++ src/typeDefs.ts
@@ -1,3 +1,4 @@
+(() => {
 const { gql } = require('apollo-server');
 
 const typeDefs = gql`
@@ -59,3 +60,4 @@ type Mutation {
 `;
 
 module.exports = typeDefs;
+})();
