+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "outDir": "./build",
+    "allowJs": true,
+    "target": "es5",
+    "types": ["node"] //need to add jest as type later on
+  },
+  "include": ["./src/**/*"]
+}
